requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
Authlib==1.6.5
banks==2.2.0
beartype==0.22.5
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
cyclopts==4.2.1
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
diskcache==5.6.3
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
exceptiongroup==1.3.0
fastmcp==2.13.0.2
filelock==3.20.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.2.4
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.6.0
llama-index-core==0.14.7
llama-index-instrumentation==0.4.2
llama-index-llms-ollama==0.9.0
llama-index-workflows==2.11.1
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mcp==1.21.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numba==0.62.1
numpy==2.3.4
ollama==0.6.0
openai-whisper==20250625
openapi-pydantic==0.5.1
packaging==25.0
pathable==0.4.4
pathvalidate==3.3.1
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.11.0
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
python-multipart==0.0.20
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.28.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
torch==2.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
wrapt==2.0.1
yarl==1.22.0
